requirements.txt
python-telegram-bot==20.3
pdfplumber
pytesseract
pdf2image
Pillow
